Rename PirDashboardActivity to PirWebViewActivity
