+++ pir/pir-impl/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
 
     <application>
         <activity
-            android:name=".dashboard.PirDashboardActivity"
+            android:name=".ui.PirWebViewActivity"
             android:parentActivityName="com.duckduckgo.app.settings.SettingsActivity"
             android:label="@string/activityPirDashboard" />
     </application>
